♻️ Question => Inquiry 변경 및 데이터 fetch 로직 추가
