--- src/components/QnaList/dummyQuestionData.json
@@ -1,86 +1,0 @@
-[
-  {
-    "questionId": 1,
-    "productId": 1,
-    "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-    "author": "ChunGu",
-    "date": "2022.02.17",
-    "answer": {
-      "questionId": 1,
-      "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-      "date": "2022.02.19"
-    }
-  },
-  {
-    "questionId": 2,
-    "productId": 1,
-    "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-    "author": "ChunGu",
-    "date": "2022.02.17",
-    "answer": {
-      "questionId": 2,
-      "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-      "date": "2022.02.19"
-    }
-  },
-  {
-    "questionId": 3,
-    "productId": 1,
-    "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-    "author": "ChunGu",
-    "date": "2022.02.17",
-    "answer": {
-      "questionId": 3,
-      "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-      "date": "2022.02.19"
-    }
-  },
-  {
-    "questionId": 4,
-    "productId": 1,
-    "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-    "author": "ChunGu",
-    "date": "2022.02.17",
-    "answer": {
-      "questionId": 4,
-      "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-      "date": "2022.02.19"
-    }
-  },
-  {
-    "questionId": 5,
-    "productId": 1,
-    "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-    "author": "ChunGu",
-    "date": "2022.02.17",
-    "answer": {
-      "questionId": 5,
-      "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-      "date": "2022.02.19"
-    }
-  },
-  {
-    "questionId": 6,
-    "productId": 1,
-    "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-    "author": "ChunGu",
-    "date": "2022.02.17",
-    "answer": {
-      "questionId": 6,
-      "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-      "date": "2022.02.19"
-    }
-  },
-  {
-    "questionId": 7,
-    "productId": 1,
-    "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-    "author": "ChunGu",
-    "date": "2022.02.17",
-    "answer": {
-      "questionId": 7,
-      "content": "감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한 감사원의 조직·직무범위·감사위원의 자격·감사대상공무원의 범위 기타 필요한",
-      "date": "2022.02.19"
-    }
-  }
-]
